Overhaul GUI to improve flexibility and high DPI support
Changes are mostly to make GUI elements display properly on devices
that have scaling applied (as is the case with many high-dpi
displays). The resize dialog was redesigned to be more streamlined.
Many small tweaks were made to some of the windows to improve
their aesthetic and consistency. Some of the GUI code was refactored
to reduce the amount of redundant code.

There is room for improvement in high DPI support, however it
relies upon GTK+3 features, which will not be available until
GIMP v3.0 is released.