Select none instead of all before calling user specified plugin
This PR originates with a ticket opened for my GIMP plugin Pixel Art Scalers:
https://github.com/bbbbbr/gimp-plugin-pixel-art-scalers/issues/15

Currently `apply_userdef()` calls `gimp_selection_all()` shortly before running the user specified plugin call. 

In most cases this is ok, however with my plugin which `resizes` the image and drawable layer, but does not `rescale` the image using libgimp calls. My plugin honors selection regions since there are cases where they might want to resize just a subset of the source image.

What this means is that the Select All area selection doesn't grow when the image gets `resized`, so the upscaled output gets cropped to the original image dimensions. When no selection area is specified then the resized drawable is able to update it's full area.

It's unclear whether there is a specific purpose for `gimp_selection_all()` or if `gimp_selection_none()` would be an acceptable alternative